baikal is work

+++ components/transport/controllers/kitAPI.js
@@ -121,37 +121,43 @@ module.exports.updateHandbookPlaces = async ctx => {
     // await fetch('https://capi.gtdel.com/1.0/geography/city/get-list?token=' + process.env.KIT)
     await fetch('https://capi.gtdel.com/1.0/tdd/city/get-list?token=' + process.env.KIT)
         .then(async response => {
-            const res = await response.json();
-            // console.log(res);
+            if (response.ok) {
+                const res = await response.json();
+                // console.log(res);
 
-            const start = Date.now();
-            let i = 0;
+                const start = Date.now();
+                let i = 0;
 
-            // очистить коллекцию населённых пунктов
-            await KitHandbookPlaces.deleteMany();
+                // очистить коллекцию населённых пунктов
+                await KitHandbookPlaces.deleteMany();
 
-            for (const city of res) {
-                //API Кита отдаёт не только города России, но также и СНГ
-                if (city.country_code !== "RU") continue;
-                console.log(city);
-                if (!(++i % 1000)) console.log('write: ', i);
-                try {
-                    await KitHandbookPlaces.create({
-                        code: city.code,
-                        name: city.name,
-                        regcode: city.region_code,
-                        requiredPickup: city.required_pickup,
-                        requiredDelivery: city.required_delivery,
-                    })
+                for (const city of res) {
+                    //API Кита отдаёт не только города России, но также и СНГ
+                    if (city.country_code !== "RU") continue;
+                    console.log(city);
+                    if (!(++i % 1000)) console.log('write: ', i);
+                    try {
+                        await KitHandbookPlaces.create({
+                            code: city.code,
+                            name: city.name,
+                            regcode: city.region_code,
+                            requiredPickup: city.required_pickup,
+                            requiredDelivery: city.required_delivery,
+                        })
+                    }
+                    catch (error) {
+                        console.log(error)
+                        continue;
+                    }
                 }
-                catch (error) {
-                    console.log(error)
-                    continue;
-                }
+
+                console.log('Kit handbook places is updated. Run time: ', ((Date.now() - start) / 1000), ' sek rows: ', i)
+                ctx.body = 'Kit handbook places is updated. Run time: ' + ((Date.now() - start) / 1000) + ' sec rows: ' + i;
             }
-
-            console.log('Kit handbook places is updated. Run time: ', ((Date.now() - start) / 1000), ' sek rows: ', i)
-            ctx.body = 'Kit handbook places is updated. Run time: ' + ((Date.now() - start) / 1000) + ' sec rows: ' + i;
+            else {
+                // console.log(await response.json());
+                throw new Error(`Error fetch query - status: ${response.status}`);
+            }
         })
         .catch(error => {
             console.log(err);
